High bit-depth bilinear

+++ src/yuv_to_rgba_bilinear.rs
@@ -76,8 +76,8 @@ fn interpolate_1_row<const DESTINATION_CHANNELS: u8, const Q: i32>(
     let g_coef_1 = transform.g_coeff_1;
     let g_coef_2 = transform.g_coeff_2;
 
-    let bias_y = range.bias_y as i32;
-    let bias_uv = range.bias_uv as i32;
+    let bias_y = range.bias_y as i16;
+    let bias_uv = range.bias_uv as i16;
 
     const BIT_DEPTH: usize = 8;
 
@@ -95,9 +95,9 @@ fn interpolate_1_row<const DESTINATION_CHANNELS: u8, const Q: i32>(
         let cb_1 = (u_src[0] as u16 + u_src[1] as u16 + 1) >> 1;
         let cr_1 = (v_src[0] as u16 + v_src[1] as u16 + 1) >> 1;
 
-        let y_value0 = (y_src[0] as i32 - bias_y) * y_coef as i32;
-        let cb_value0 = cb_0 as i16 - bias_uv as i16;
-        let cr_value0 = cr_0 as i16 - bias_uv as i16;
+        let y_value0 = (y_src[0] as i32 - bias_y as i32) * y_coef as i32;
+        let cb_value0 = cb_0 as i16 - bias_uv;
+        let cr_value0 = cr_0 as i16 - bias_uv;
 
         let r0 = qrshr::<Q, BIT_DEPTH>(y_value0 + cr_coef as i32 * cr_value0 as i32);
         let b0 = qrshr::<Q, BIT_DEPTH>(y_value0 + cb_coef as i32 * cb_value0 as i32);
@@ -114,9 +114,9 @@ fn interpolate_1_row<const DESTINATION_CHANNELS: u8, const Q: i32>(
             rgba0[dst_chans.get_a_channel_offset()] = 255u8;
         }
 
-        let y_value1 = (y_src[1] as i32 - bias_y) * y_coef as i32;
-        let cb_value1 = cb_1 as i16 - bias_uv as i16;
-        let cr_value1 = cr_1 as i16 - bias_uv as i16;
+        let y_value1 = (y_src[1] as i32 - bias_y as i32) * y_coef as i32;
+        let cb_value1 = cb_1 as i16 - bias_uv;
+        let cr_value1 = cr_1 as i16 - bias_uv;
 
         let r0 = qrshr::<Q, BIT_DEPTH>(y_value1 + cr_coef as i32 * cr_value1 as i32);
         let b0 = qrshr::<Q, BIT_DEPTH>(y_value1 + cb_coef as i32 * cb_value1 as i32);
@@ -140,15 +140,15 @@ fn interpolate_1_row<const DESTINATION_CHANNELS: u8, const Q: i32>(
     let rgba_remainder = rgba_chunks.into_remainder();
 
     if let ([last_y], rgba) = (y_remainder, rgba_remainder) {
-        let y_value0 = (*last_y as i32 - bias_y) * y_coef as i32;
-        let cb_value = *u_plane.last().unwrap() as i32 - bias_uv;
-        let cr_value = *v_plane.last().unwrap() as i32 - bias_uv;
+        let y_value0 = (*last_y as i32 - bias_y as i32) * y_coef as i32;
+        let cb_value = *u_plane.last().unwrap() as i16 - bias_uv;
+        let cr_value = *v_plane.last().unwrap() as i16 - bias_uv;
         let rgba0 = &mut rgba[..channels];
 
-        let r0 = qrshr::<Q, BIT_DEPTH>(y_value0 + cr_coef as i32 * cr_value);
-        let b0 = qrshr::<Q, BIT_DEPTH>(y_value0 + cb_coef as i32 * cb_value);
+        let r0 = qrshr::<Q, BIT_DEPTH>(y_value0 + cr_coef as i32 * cr_value as i32);
+        let b0 = qrshr::<Q, BIT_DEPTH>(y_value0 + cb_coef as i32 * cb_value as i32);
         let g0 = qrshr::<Q, BIT_DEPTH>(
-            y_value0 - g_coef_1 as i32 * cr_value - g_coef_2 as i32 * cb_value,
+            y_value0 - g_coef_1 as i32 * cr_value as i32 - g_coef_2 as i32 * cb_value as i32,
         );
         rgba0[dst_chans.get_r_channel_offset()] = r0 as u8;
         rgba0[dst_chans.get_g_channel_offset()] = g0 as u8;
@@ -180,8 +180,8 @@ fn interpolate_2_rows<const DESTINATION_CHANNELS: u8, const Q: i32>(
     let g_coef_1 = transform.g_coeff_1;
     let g_coef_2 = transform.g_coeff_2;
 
-    let bias_y = range.bias_y as i32;
-    let bias_uv = range.bias_uv as i32;
+    let bias_y = range.bias_y as i16;
+    let bias_uv = range.bias_uv as i16;
 
     const BIT_DEPTH: usize = 8;
 
@@ -221,9 +221,9 @@ fn interpolate_2_rows<const DESTINATION_CHANNELS: u8, const Q: i32>(
             + (1 << 3))
             >> 4;
 
-        let y_value0 = (y_src0[0] as i32 - bias_y) * y_coef as i32;
-        let cb_value0 = cb_0 as i16 - bias_uv as i16;
-        let cr_value0 = cr_0 as i16 - bias_uv as i16;
+        let y_value0 = (y_src0[0] as i32 - bias_y as i32) * y_coef as i32;
+        let cb_value0 = cb_0 as i16 - bias_uv;
+        let cr_value0 = cr_0 as i16 - bias_uv;
 
         let g_built_coeff0 =
             -g_coef_1 as i32 * cr_value0 as i32 - g_coef_2 as i32 * cb_value0 as i32;
@@ -241,9 +241,9 @@ fn interpolate_2_rows<const DESTINATION_CHANNELS: u8, const Q: i32>(
             rgba00[dst_chans.get_a_channel_offset()] = 255u8;
         }
 
-        let y_value1 = (y_src0[1] as i32 - bias_y) * y_coef as i32;
-        let cb_value1 = cb_1 as i16 - bias_uv as i16;
-        let cr_value1 = cr_1 as i16 - bias_uv as i16;
+        let y_value1 = (y_src0[1] as i32 - bias_y as i32) * y_coef as i32;
+        let cb_value1 = cb_1 as i16 - bias_uv;
+        let cr_value1 = cr_1 as i16 - bias_uv;
 
         let g_built_coeff1 =
             -g_coef_1 as i32 * cr_value1 as i32 - g_coef_2 as i32 * cb_value1 as i32;
@@ -268,21 +268,21 @@ fn interpolate_2_rows<const DESTINATION_CHANNELS: u8, const Q: i32>(
     let rgba_remainder = rgba_chunks.into_remainder();
 
     if let ([last_y], rgba) = (y_remainder, rgba_remainder) {
-        let y_value0 = (*last_y as i32 - bias_y) * y_coef as i32;
+        let y_value0 = (*last_y as i32 - bias_y as i32) * y_coef as i32;
 
         let cb_0 =
             (*u_plane0.last().unwrap() as u16 * 3 + *u_plane1.last().unwrap() as u16 + (1 << 1))
                 >> 2;
         let cr_0 = (*v_plane0.last().unwrap() as u16 + (*v_plane1.last().unwrap()) as u16 + 1) >> 1;
 
-        let cb_value = cb_0 as i32 - bias_uv;
-        let cr_value = cr_0 as i32 - bias_uv;
+        let cb_value = cb_0 as i16 - bias_uv;
+        let cr_value = cr_0 as i16 - bias_uv;
         let rgba0 = &mut rgba[..channels];
 
-        let g_built_coeff = -g_coef_1 as i32 * cr_value - g_coef_2 as i32 * cb_value;
+        let g_built_coeff = -g_coef_1 as i32 * cr_value as i32 - g_coef_2 as i32 * cb_value as i32;
 
-        let r0 = qrshr::<Q, BIT_DEPTH>(y_value0 + cr_coef as i32 * cr_value);
-        let b0 = qrshr::<Q, BIT_DEPTH>(y_value0 + cb_coef as i32 * cb_value);
+        let r0 = qrshr::<Q, BIT_DEPTH>(y_value0 + cr_coef as i32 * cr_value as i32);
+        let b0 = qrshr::<Q, BIT_DEPTH>(y_value0 + cb_coef as i32 * cb_value as i32);
         let g0 = qrshr::<Q, BIT_DEPTH>(y_value0 + g_built_coeff);
 
         rgba0[dst_chans.get_r_channel_offset()] = r0 as u8;
